WL-3576 : Add email sent confirmation message to feedback tool
This is only for logged in users.

For the Content Report, I'm grabbing the siteUpdater in the js and showing it in a box at the top.
For the Technical Report, same thing except that the email address is normally got in Java after the submission so I'm using the same logic before to put it in the request before getting it out again in the js afterwards.