WL-3592 : contact list is empty
This is because at the moment the logic in the tool is to only show the site maintainers if the site has no email but this requirement has changed.
Now they are shown if the user is logged in and has the right permission (the permission bit is in another JIRA ; for now, just show the site maintainers if the user is logged in). .