WL-3623 : List of maintainers is blank if site contains a deleted user
The problem is that if you go to the Contact Us tool for a site that has
a (maintainer/contributor) user that has been deleted, then the site gets
its users (including the deleted one) and tries to look up information on
that user, throws a UserNotDefinedException, and returns an empty list.
This means that a site with one deleted user doesn't show
any of the other non-deleted maintainers in the Contact Us tool.

This fix makes use of userDirectoryService.getUsers() to get the
users - which gets the users from the db and does not throw an exception
if it can't find a particualr user.