WL-3819: EntityProviders do not like forward slashes in their post urls
If someone typed in a non-existent site like:

https://staging-werp.bsp.ox.ac.uk/portal/hierarchy/central/oucs/test/ada

and went to the Contact Us tool, then we create a form with a post action url of:

/direct/feedback/central/oucs/test/ada/reportcontent

which the EntityProvider backing the form does not like because it parses the url by forward slashes - which you can see from:

https://staging-werp.bsp.ox.ac.uk/direct/feedback/describe

So I'm replacing the forward slashes, if there are any, before and after the submission (which is ugly).

Using any of the methods of StringEscapeUtils or escaping the forward slashes does not work.